WORK

+++ DagraacSystems/Scripts/Network/ISession.cs
@@ -4,7 +4,7 @@
 	{
 		bool IsConnected { get; }
 
-		void Connect(string ip, ushort port);
+		void Connect(string ip, int port);
 		void Disconnect();
 		void Send(byte[] data);
 	}
